Migrate tests across the vix/vixen split; share decode_request in core
- Move decode's document→PrimitiveValue logic into vix-core
  (runtime::decode_request); vixen's DecodePrimitive is now a thin RawPrimitive
  wrapper over it. The scheduler's own decode-dispatch unit tests wrap the same
  in-crate decode_request via a LocalDecode double — a dev-dependency on the real
  DecodePrimitive would be a second copy of vix-core (the dev-dep cycle), whose
  RawPrimitive would not satisfy this crate's own trait.
- vix-core dev-depends on the vixen crates so its integration tests can drive the
  runnable system (ratchet/budget → vixen_runtime::; PRELUDE_SOURCES is &[&str]
  data, so naming it across the cycle is identity-hazard-free).
- The two budget-child-bin tests move to vixen-runtime/tests/rung050_trace_budget.rs
  (CARGO_BIN_EXE_vix-budget-child is only nameable from that crate); their ratchet
  fixtures stay in vix-core/tests/ratchet and are referenced by relative path.

All four crates compile with tests. Test-run verification follows.
